requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.6
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyMatting==1.1.13
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
